refs #2489 specify camera host and frame path to be processed and refs #2334 option parsing

+++ leginon/rawtransfer.py
@@ -3,6 +3,7 @@
 import os
 import sys
 import shutil
+from optparse import OptionParser
 import subprocess
 import time
 import leginon.leginondata
@@ -11,14 +12,55 @@ import pyami.fileutil
 
 check_interval = 20  # seconds between checking for new frames
 
-def get_source_path():
-	raw_frames_path = 'RAW_FRAMES_PATH'
-	try:
-		source_path = os.environ[raw_frames_path]
-	except:
-		sys.stderr.write('specify location of raw frames in variable %s\n' % (raw_frames_path,))
+def parseParams():
+	'''
+	Use OptionParser to get parameters
+	'''
+	parser = OptionParser()
+
+	# options
+	parser.add_option("--method", dest="method",
+		help="method to transfer, e.g. --method=mv", type="choice", choices=['mv','rsync'], default='rsync')
+	parser.add_option("--source_path", dest="source_path",
+		help="Mounted parent path to transfer, e.g. --source_path=/mnt/ddframes", metavar="PATH")
+	parser.add_option("--camera_host", dest="camera_host",
+		help="Camera computer hostname in leginondb, e.g. --camera_host=gatank2")
+	parser.add_option("--destination_head", dest="dest_path_head",
+		help="Specific head destination frame path to transfer if multiple frame transfer is run for one source to frame paths not all mounted on the same computer, e.g. --destination_head=/data1", metavar="PATH", default='')
+
+	# parsing options
+	(options, optargs) = parser.parse_args(sys.argv[1:])
+	if len(optargs) > 0:
+		print "Unknown commandline options: "+str(optargs)
+	if len(sys.argv) < 2:
+		parser.print_help()
+		parser.error("no options defined")
+	params = {}
+	for i in parser.option_list:
+		if isinstance(i.dest, str):
+			params[i.dest] = getattr(options, i.dest)
+	checkOptionConflicts(params)
+	return params
+
+def checkOptionConflicts(params):
+	if params['camera_host']:
+		r = leginon.leginondata.InstrumentData(hostname=params['camera_host']).query()
+		if not r:
+			sys.stderr.write('Camera host %s not in Leginon database\n' % (params['camera_host']))
+			sys.exit(1)
+
+def getAndValidatePath(params,key):
+	pathvalue = params[key]
+	if pathvalue and not os.access(pathvalue, os.R_OK):
+		sys.stderr.write('%s not exists or not readable\n' % (pathvalue,))
 		sys.exit(1)
-	return source_path
+	return pathvalue
+
+def get_source_path(params):
+	return getAndValidatePath(params,'source_path')
+
+def get_dst_head(params):
+	return getAndValidatePath(params,'dest_path_head')
 
 image_classes = [
 	leginon.leginondata.AcquisitionImageData,
@@ -27,8 +69,9 @@ image_classes = [
 	leginon.leginondata.NormImageData,
 ]
 
-def query_image_by_frames_name(name):
-	qcam = leginon.leginondata.CameraEMData()
+def query_image_by_frames_name(name,cam_host):
+	qccd = leginon.leginondata.InstrumentData(hostname=cam_host)
+	qcam = leginon.leginondata.CameraEMData(ccdcamera=qccd)
 	qcam['frames name'] = name
 	for cls in image_classes:
 		qim = cls(camera=qcam)
@@ -37,17 +80,45 @@ def query_image_by_frames_name(name):
 			return results[0]
 	return None
 
-def copy_and_delete(src, dst, uid, gid):
+def copy_and_delete(src, dst):
+	'''
+	Use rsync to copy the file.  The sent files are removed
+	after copying.
+	'''
+	cmd = 'rsync -av --remove-sent-files %s %s' % (src, dst)
+	print cmd
+	p = subprocess.Popen(cmd, shell=True)
+	p.wait()
+
+def move(src, dst):
+	'''
+	Use mv command to rename frames
+	'''
+	cmd = 'mv %s %s' % (src, dst)
+	print cmd
+	p = subprocess.Popen(cmd, shell=True)
+	p.wait()
+
+def transfer(src, dst, uid, gid, method):
+	'''
+	This function at minimal organize and rename the time-stamped file
+	to match the Leginon session and integrated image.  If the source is
+	saved on the local drive of the camera computer, rsync (default) is
+  used to copy the file off and the time-stamped file on the camera
+	local drive removed to make room for more data collection.
+	'''
 	# make destination dirs
 	dirname,basename = os.path.split(os.path.abspath(dst))
 	print 'mkdirs', dirname
 	pyami.fileutil.mkdirs(dirname)
 
-	# copy frames
-	cmd = 'rsync -av --remove-sent-files %s %s' % (src, dst)
-	print cmd
-	p = subprocess.Popen(cmd, shell=True)
-	p.wait()
+	if method == 'rsync':
+		# safer method but slower
+		copy_and_delete(src, dst)
+	else:
+		# move does only rename and therefore will be faster if on the same device
+		# but could have problem if dropped during the process between devices.
+		move(src,dst)
 
 	# change ownership of desintation directory and contents
 	cmd = 'chown -R %s:%s %s' % (uid, gid, dirname)
@@ -55,22 +126,23 @@ def copy_and_delete(src, dst, uid, gid):
 	p = subprocess.Popen(cmd, shell=True)
 	p.wait()
 
-	# remove empty dir from source
-	abspath = os.path.abspath(src)
-	dirpath,basename = os.path.split(abspath)
-	if src.endswith('/'):
-		cmd = 'find %s -type d -empty -prune -exec rmdir --ignore-fail-on-non-empty -p \{\} \;' % (basename,)
-	print 'cd', dirpath
-	print cmd
-	p = subprocess.Popen(cmd, shell=True, cwd=dirpath)
-	p.wait()
+	if method == 'rsync':
+		# remove empty .frames dir from source
+		abspath = os.path.abspath(src)
+		dirpath,basename = os.path.split(abspath)
+		if src.endswith('/'):
+			cmd = 'find %s -type d -empty -prune -exec rmdir --ignore-fail-on-non-empty -p \{\} \;' % (basename,)
+			print 'cd', dirpath
+		print cmd
+		p = subprocess.Popen(cmd, shell=True, cwd=dirpath)
+		p.wait()
 
 next_time_start = 0
 mtime = 0
 time_expire = 300  # ignore anything older than 5 minutes
 expired_names = {} # directories that have no db record after 5 minutes
 
-def run_once(parent_src_path):
+def run_once(parent_src_path,cam_host,dest_head,method):
 	global next_time_start
 	global time_expire
 	global mtime
@@ -92,6 +164,9 @@ def run_once(parent_src_path):
 		if mtime > next_time_start:
 			next_time_start = mtime
 
+		# ignore irrelevent source files or folders
+		if not name.endswith('.mrc') and not name.endswith('.frames'):
+			continue
 		print '**running', src_path
 
 		# query for Leginon image
@@ -104,14 +179,20 @@ def run_once(parent_src_path):
 			## ensure a trailing / on directory
 			if src_path[-1] != '/':
 				src_path = src_path + '/'
-		imdata = query_image_by_frames_name(frames_name)
+		imdata = query_image_by_frames_name(frames_name,cam_host)
 		if imdata is None:
 			continue
 		image_path = imdata['session']['image path']
 		frames_path = imdata['session']['frame path']
 
+		# back compatible to sessions without frame path in database
 		if image_path and not frames_path:
 			frames_path = ddinfo.getRawFrameSessionPathFromImagePath(image_path)
+
+		# only process destination frames_path starting with chosen head
+		if not frames_path.startswith(dest_head):
+			print '    Destination frame path does not starts with %s. Skipped' % (dest_head)
+			continue
 
 		# determine user and group of leginon data
 		stat = os.stat(image_path)
@@ -120,18 +201,23 @@ def run_once(parent_src_path):
 
 		imname = imdata['filename'] + dst_suffix
 		dst_path = os.path.join(frames_path, imname)
-		print 'DEST', dst_path
+		print 'Destination path: %s' %  (dst_path)
 
 		# do actual copy and delete
-		copy_and_delete(src_path, dst_path, uid, gid)
+		transfer(src_path, dst_path, uid, gid,method)
 		leginon.ddinfo.saveImageDDinfoToDatabase(imdata,os.path.join(dst_path,'info.txt'))
 
 def run():
-	src_path = get_source_path()
+	params = parseParams()
+	src_path = get_source_path(params)
 	print 'Source path:  %s' % (src_path,)
+	dst_head = get_dst_head(params)
+	if dst_head:
+		print "Limit processing to destination frame path started with %s" % (dst_head)
 	while True:
 		print 'Iterating...'
-		run_once(src_path)
+		run_once(src_path,params['camera_host'],dst_head,method=params['method'])
+		print 'Sleeping...'
 		time.sleep(check_interval)
 
 if __name__ == '__main__':
